pyproject.toml
[project]
name = "adh-cli"
version = "0.1.0"
description = "A TUI application using Textual and Google ADK"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "textual>=0.47.0",
    "google-genai>=0.1.0",
    "rich>=13.7.0",
    "click>=8.1.7",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "textual-dev>=1.3.0",
]

[project.scripts]
adh-cli = "adh_cli.__main__:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["adh_cli"]